feat(permissions): role permission based permission module switching
To facilitate a gradual migration to the permissions module, introduce
a role permission based check, to evaluate whether the permission module
should apply. In addition, a new permission switcher method is added,
that calls the registered new variant of the method in question.
With this change, permission module logic can be applied, when the
current user's role permission has been migrated and should adhere
to the permissions module, without having to duplicate the permissions
module based logic in the RBAC logic.