feat(docs): migration tooling for migrating to s3
This is intended to synchronize the documents module's data from the
filesystem storage (NFS or local) to S3. It uses a specialized tool to do
this, but adds code to verify the migrated data is complete and correct.

Note: This is intended for dev env usage only. The production migration will
be run from storage to storage directly, not from within a django container.