fix(be): use date configuration for municipality deactivation
Use a date configuration on the service meta to determine whether
the responsible municipality on an instance is deactivated (if the
current date/time is past the configured date/time, then the
municipality is considered deactivated). In addition, ensure that
only active instance services of service group "municipality" are
taken into account.