refactor(permissions): use internal context object instead of bare instance
This allows module-specific permission managers to extend the context
to add more information, which in turn can be used by custom permission
conditions to check additional things.

For example, in the document module (alexandria), a custom permission
manager can extend the context to add a document or category, which then
can be used to determine a permission on said document.

Co-authored-by: Jonas Metzener <jonas.metzener@adfinis.com>