feat(alexandria): implement new permission module adaptation
**Add new module-specific permission manager:**

This permissions manager - together with it's additional custom checks -
can be used to leverage the permission module's infrastructure to
implement it's own permissions system, using the same patterns as the
global permissions module.

**Add new DGAP permission class for alexandria:**

Based on the module specific permission manager, introduce a DGAP
permission class.

**Add permission scopes:**

The new permission scope is the combination of a permission manager and
a permission context. The manager and context may be module-specific.

This allows us to create scoped permission checks that can differ between
one document and another.

Co-authored-by: Jonas Metzener <jonas.metzener@adfinis.com>