chore(alexandria): temporarily allow all actions in v2 permissions
Since the older cantons that will use the new v2 permission class have
not fully integrated the permission module yet, we allow all alexandria
actions so we can test the integration of alexandria.

This will need to be removed as soon as the permission module is fully
integrated and the alexandria permissions are configured.