feat(tests): introduce ability to define sql views
For performance reasons, we build the test databases by having pytest-django
scan the models and build the DB directly, instead of running the migrations
one by one.

This breaks down if we have migrations that define SQL views, as those won't
be created that way.

However, SQL views are very useful in gathering information and represent it
in a way that might be cumbersome to do with just the Django ORM.

Therefore, you can now register migrations that define SQL views, and
those views will then still be created in the test databases.