fix(billing): initialise product number for new records
The selected product number was only set upon changing the selection,
but that is unlikely to happen (we set a default, and there's usually
only one option available anyway). So submission would usually fail due
to a missing product number on the object itself. This change ensures we
have a product number set on a newly created entry.

We also refactor the product number fetching a bit: Rather than fetching
the product numbers asynchronously (which makes setting the
initial/default selection difficult), we do so as part of creating the
record.

An ugly artefact of that is that we set a controller variable
(this.productNumbers) as part of creating a record. We could have
avoided this by setting the entry's product number as part of fetching
the product number instead, but that is just ugly in another way, and
would have rendered the code less understandable.