fix(ag): remove work items with template-id NONE
This is a leftover from !10621 where manual work items not created from
a template needed to be marked with template-id "NONE" in order for the
filters to work. This is now obsolete as the filtering is done on the
dedicated REST API endpoint for the work item list - not Caluma anymore.