refactor(placholders): add support for collections and decouple code
The field docs and available placholder bit was pretty hard to grasp
(unless you need not touch it, then it's probably fine.)

Every placeholder field inherits from AliasedMixin (might be renamed
to better capture what it does). Every placeholder needs documenting and
therefore that's probably the best place to produce the documenting.

 - let the field produce its docs
 - add more comments
 - keep concerns better separated (split to methods and don't name
   variables alias that are object keys etc. .. so confusing.)
 - only add those placeholder translations that acutally make sense.
   E. g.: don't add nested keys that mix languages in their dot-path

NOTE: the collection of placeholders in case of collection or lists of objects
will retain an entry for

 - alias_translation
 - alias_translation[]

The first does not appear necessary but is nevertheless added to maintain
consistency with the existing snapshots.