feat(sz): allow reader role to send communication messages
Technically the 'is reader?' check is frontend-only, so this is mainly a
cosmetic change.

We could extend this fix (adding backend checks and/or adding a feature
flag to limit this behaviour to SZ only in a cleaner way), but since
this is subject to being covered by the upcoming permissions module
anyway, this fairly low-effort temporary workaround should do for now.

To be cleaned up when the permissions module takes full effect on the
older cantons.