chore(compose/sz): limit port mapping for Metabase to dev environment
Metabase is accessible as /metabase/, so generally it shouldn't need a
separate port to access it directly.

On stage/prod, the port isn't allowed from the outside, so we can remove
it altogether.

On the local dev environment, however, the Nginx reverse proxy adds CSP
headers that currently cause issues, so as a temporary workaround is to
use the port mapping. To avoid issues with rootless setups, pick a port
number that is not in the privileged range (because 3000 is used by
Unoconv already, we go with 3001).